+++ src/Components/BestDeals/BestDeals.module.css
@@ -8,3 +8,9 @@
 a {
     text-decoration: none;
 }
+
+@media (min-width: 480px) {
+    .BestDeals {
+        max-width: 100%;
+    }
+}
