+++ src/Components/Cards/BigCard/BigCard.module.css
@@ -54,3 +54,9 @@
 .Card .CustomButton {
     margin-top: 15px;
 }
+
+@media (min-width: 480px) {
+    .Card {
+        width: 450px;
+    }
+}
