+++ src/Components/Filters/FilterItems/FilterItems.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import './FilterItems.css';
+
+const FilterItems = props => {
+
+    return (
+        <>
+
+        </>
+    );
+};
+
+export default FilterItems;
