+++ src/Components/Filters/ProductFilter.css
@@ -1,0 +1,30 @@
+.filter-button {
+    display: flex;
+}
+
+.filter-card {
+    display: none;
+}
+
+@media (min-width: 480px) {
+    .filter-button {
+        display: none;
+    }
+
+    .filter-card {
+        display: block;
+        width: 300px;
+        max-width: 100%;
+        height: 400px;
+        max-height: fit-content;
+        border: 1px solid white;
+        background-color: white;
+        border-radius: 2%;
+        padding: 16px 10px;
+        margin: 16px 10px;
+        box-sizing: border-box;
+        text-align: center;
+        overflow: hidden;
+        color: black;
+    }
+}
