--- src/Components/ProductFilter/ProductFilter.css
@@ -1,4 +1,0 @@
-/* .FontAwesomeIcon {
-    height: 14px;
-    color: black;
-} */
