+++ src/Services/ProductsService.js
@@ -2,11 +2,6 @@ import { firebaseDb } from '../Config/firebaseConfig';
 
 const docRef = firebaseDb.collection('/products');
 
-const PRODUCTS = 'products';
-const NEWRELEASES = 'newReleases';
-const BESTDEALS = 'bestDeals'
-const TOPSELLERS = 'topSellers'
-
 const getById = (id) =>
     docRef.doc(id)
         .get()
@@ -15,7 +10,7 @@ const getById = (id) =>
 const getByCategory = (category) => 
     docRef.where('category', '==', category)
         .get()
-        .then(res => res.docs.map(doc => doc.data()))
+        .then(res => res.docs.map(doc => doc.data()));
 
 // const getByItemType = (itemType) => db.child(PRODUCTS).orderByChild('itemType').equalTo(itemType);
